use an Int for ZIO parallelism combinators

+++ core/shared/src/main/scala/zio/IO.scala
@@ -85,7 +85,7 @@ object IO {
   /**
    * @see See [[zio.ZIO.collectAllParN]]
    */
-  final def collectAllParN[E, A](n: Long)(as: Iterable[IO[E, A]]): IO[E, List[A]] =
+  final def collectAllParN[E, A](n: Int)(as: Iterable[IO[E, A]]): IO[E, List[A]] =
     ZIO.collectAllParN(n)(as)
 
   /**
@@ -188,7 +188,7 @@ object IO {
   /**
    * @see See [[zio.ZIO.foreachParN]]
    */
-  final def foreachParN[E, A, B](n: Long)(as: Iterable[A])(fn: A => IO[E, B]): IO[E, List[B]] =
+  final def foreachParN[E, A, B](n: Int)(as: Iterable[A])(fn: A => IO[E, B]): IO[E, List[B]] =
     ZIO.foreachParN(n)(as)(fn)
 
   /**
@@ -206,7 +206,7 @@ object IO {
   /**
    * @see See [[zio.ZIO.foreachParN_]]
    */
-  final def foreachParN_[E, A, B](n: Long)(as: Iterable[A])(f: A => IO[E, _]): IO[E, Unit] =
+  final def foreachParN_[E, A, B](n: Int)(as: Iterable[A])(f: A => IO[E, _]): IO[E, Unit] =
     ZIO.foreachParN_(n)(as)(f)
 
   /**
@@ -377,7 +377,7 @@ object IO {
   /**
    *  See [[zio.ZIO.sequenceParN]]
    */
-  final def sequenceParN[E, A](n: Long)(as: Iterable[IO[E, A]]): IO[E, List[A]] =
+  final def sequenceParN[E, A](n: Int)(as: Iterable[IO[E, A]]): IO[E, List[A]] =
     ZIO.sequenceParN(n)(as)
 
   /**
@@ -457,7 +457,7 @@ object IO {
    * Alias for [[ZIO.foreachParN]]
    */
   final def traverseParN[E, A, B](
-    n: Long
+    n: Int
   )(as: Iterable[A])(fn: A => IO[E, B]): IO[E, List[B]] =
     ZIO.traverseParN(n)(as)(fn)
 
@@ -477,7 +477,7 @@ object IO {
    * @see See [[zio.ZIO.traverseParN_]]
    */
   final def traverseParN_[E, A](
-    n: Long
+    n: Int
   )(as: Iterable[A])(f: A => IO[E, _]): IO[E, Unit] =
     ZIO.traverseParN_(n)(as)(f)
